Size bars from the full time signature in rulers, grids and readouts
A bar is numerator times the denominator's note in quarter-note beats, and
the displayed beat is the signature's beat, so 6/8 draws three-quarter bars
counted in eighths and the auto grid subdivides the signature's beat (#2802).

+++ magda/daw/core/ChordProgressionConverter.hpp
@@ -30,7 +30,7 @@ struct ExtractedChord {
  * in-editor "detect chords" action and the "extract to chord track" feature.
  */
 std::vector<ExtractedChord> extractChordsFromNotes(const std::vector<MidiNote>& notes,
-                                                   int beatsPerBar);
+                                                   double beatsPerBar);
 
 /**
  * @brief Build a canonical root-position voicing for a chord.
